moved inspect_string_custom_omit_overlap into separate file
